requirements.txt
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.0.12
decorator==4.4.2
idna==3.3
imageio==2.26.0
imageio-ffmpeg==0.4.8
loguru==0.6.0
moviepy==1.0.3
mutagen==1.46.0
numpy==1.24.2
proglog==0.1.10
pycryptodomex==3.17
pyparsing==3.0.9
pyTelegramBotAPI==4.10.0
python-dotenv==0.20.0
requests==2.28.0
tqdm==4.64.1
urllib3==1.26.14
vk-api==11.9.8
websockets==10.4
youtube-dl==2021.12.17
yt-dlp==2022.7.18
